requirements.txt
protobuf==3.12.4
pyzmq==23.2.1
pytz==2023.3.post1
numpy==1.24.4
pandas==2.0.3
pyalgotrade==0.20
aioredis==1.3.1
redis==5.0.1
pymongo==4.5.0
motor==3.3.1
orjson==3.6.1